requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
sentence-transformers==2.2.2
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
python-docx==0.8.11
PyPDF2==3.0.1
pillow>=10.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
google-generativeai>=0.3.0
transformers>=4.35.0
torch>=2.1.0
requests>=2.31.0
aiofiles>=23.2.0
websockets>=12.0
